replaced webm encoder with ogg theora
git-svn-id: http://crashrpt.googlecode.com/svn@1410 821a4644-a21a-0410-aec7-6199d47568cc

+++ branches/video_record/reporting/crashsender/smtpclient.cpp
@@ -869,7 +869,7 @@ int CSmtpClient::Base64EncodeAttachment(CString sFileName,
         return 1;  // File not found.
 
     // Allocate buffer of file size
-    uFileSize = st.st_size;
+    uFileSize = (int)st.st_size;
     uchFileData = new BYTE[uFileSize];
 
     // Read file data to buffer.
